- package name - android target version - translation italy
