+++ app/build.gradle
@@ -10,7 +10,7 @@ android {
 
     defaultConfig {
         applicationId "com.app.revyou"
-        minSdkVersion 25
+        minSdkVersion 23
         targetSdkVersion 29
         versionCode 1
         versionName "1.0"
