+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.app.revyou">
+    package="com.rievoluzione.revyou">
 
     <application
         android:allowBackup="true"
@@ -9,16 +9,17 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activities.AperturaActivity">
+        <activity android:name="com.rievoluzione.revyou.activities.AperturaActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activities.PrimaActivity"></activity>
-        <activity android:name=".dialogs.DialogAcced"></activity>
-        <activity android:name=".activities.Main"></activity>
-        <activity android:name=".activities.MainActivity"></activity>
+        <activity android:name="com.rievoluzione.revyou.activities.MainActivity"></activity>
+        <activity android:name="com.rievoluzione.revyou.activities.PrimaActivity"></activity>
+        <activity android:name="com.rievoluzione.revyou.dialogs.DialogAcced"></activity>
+        <activity android:name="com.rievoluzione.revyou.activities.Main"></activity>
+
     </application>
 
 </manifest>
