--- app/src/main/java/com/app/revyou/adapter/PreferitiAdapter.java
@@ -1,60 +1,0 @@
-package com.app.revyou.adapter;
-
-import android.content.Context;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-
-import com.app.revyou.R;
-
-import java.util.List;
-
-import butterknife.BindView;
-import butterknife.ButterKnife;
-
-public class PreferitiAdapter extends RecyclerView.Adapter<PreferitiAdapter.PreferitiViewHolder> {
-
-
-    private Context mContext;
-
-    public PreferitiAdapter(Context context){
-        this.mContext =context;
-
-    }
-
-
-    @NonNull
-    @Override
-    public PreferitiViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.single_list_preferiti, parent, false);
-        PreferitiAdapter.PreferitiViewHolder holder = new PreferitiAdapter.PreferitiViewHolder(view);
-        return holder;
-
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull PreferitiViewHolder holder, int position) {
-
-    }
-
-    @Override
-    public int getItemCount() {
-        return 10;
-    }
-
-    public class PreferitiViewHolder extends RecyclerView.ViewHolder {
-
-
-
-        public PreferitiViewHolder(View itemView) {
-            super(itemView);
-            ButterKnife.bind(this, itemView);
-        }
-    }
-}
