--- app/src/main/java/com/app/revyou/dialogs/DialogAcced.java
@@ -1,16 +1,0 @@
-package com.app.revyou.dialogs;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-import com.app.revyou.R;
-
-public class DialogAcced extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_dialog_acced);
-    }
-}
