--- app/src/main/java/com/app/revyou/fragments/LoginFragment.java
@@ -1,4 +1,0 @@
-package com.app.revyou.fragments;
-
-public class LoginFragment {
-}
