--- app/src/main/java/com/app/revyou/fragments/ProfilioFragment.java
@@ -1,24 +1,0 @@
-package com.app.revyou.fragments;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-
-import com.app.revyou.R;
-
-
-public class ProfilioFragment extends androidx.fragment.app.Fragment {
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-
-        View view = inflater.inflate(R.layout.frgament_profilio, container, false);
-        return view;
-
-    }
-}
