--- app/src/main/java/com/app/revyou/fragments/RegistrazioneFragment.java
@@ -1,5 +1,0 @@
-package com.app.revyou.fragments;
-
-public class RegistrazioneFragment {
-
-}
