+++ app/src/main/java/com/rievoluzione/revyou/dialogs/DialogAcced.java
@@ -1,0 +1,16 @@
+package com.rievoluzione.revyou.dialogs;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.rievoluzione.revyou.R;
+
+public class DialogAcced extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_dialog_acced);
+    }
+}
