+++ app/src/main/java/com/rievoluzione/revyou/fragments/LoginFragment.java
@@ -1,0 +1,4 @@
+package com.rievoluzione.revyou.fragments;
+
+public class LoginFragment {
+}
