+++ app/src/main/java/com/rievoluzione/revyou/fragments/RegistrazioneFragment.java
@@ -1,0 +1,5 @@
+package com.rievoluzione.revyou.fragments;
+
+public class RegistrazioneFragment {
+
+}
