+++ app/src/main/res/layout/activity_auth.xml
@@ -44,7 +44,7 @@
                     <FrameLayout
                         android:layout_width="match_parent"
                         android:layout_height="400dp">-->
-                <com.app.revyou.activities.NonSwipeableViewPager
+                <com.rievoluzione.revyou.activities.NonSwipeableViewPager
                     android:id="@+id/view_pager"
                     android:layout_marginTop="@dimen/margin_20dp"
                     android:layout_width="match_parent"
