+++ app/src/main/res/layout/main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/lyt_parent"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.app.revyou.activities.Main">
+    tools:context="com.rievoluzione.revyou.activities.Main">
 
     <RelativeLayout
