+++ app/src/main/res/values-it-rIT/strings.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="app_name">RevYou</string>
+    <string name="Pagamento">Pagamento in unica soluzione \nda 179,88€</string>
+    <string name="Abbonamento">Abbonamento annuale</string>
+    <string name="Start">Entra in RevYou</string>
+</resources>
